changement

+++ Code/InfiniteBattlefront/app/src/main/java/com/example/larse1432921/infinitebattlefront/Turret.kt
@@ -5,10 +5,10 @@ import java.util.*
 
 var enemyStack: Stack<Enemy>? = null
 
-open class Turret(_type: Char, _x: Float, _y: Float, _range: Float, _towerView: ImageView) {
+open class Turret(_type: Char, _x: Int, _y: Int, _range: Float, _towerView: ImageView) {
     var type: Char = _type
-    var x: Float = _x
-    var y: Float = _y
+    var x: Int = _x
+    var y: Int = _y
     var range: Float = _range
     var towerView: ImageView = _towerView
     val _enemyStack: Stack<Enemy>? = enemyStack
